fix: move types to one central place the types dir
