+++ packages/esm-active-visits-app/src/visits-summary/visit-detail.component.tsx
@@ -2,8 +2,9 @@ import React, { useMemo, useState } from 'react';
 import classNames from 'classnames';
 import { useTranslation } from 'react-i18next';
 import { ContentSwitcher, DataTableSkeleton, Switch } from '@carbon/react';
-import { type Encounter, useVisit } from './visit.resource';
-import { formatTime, formatDatetime, parseDate } from '@openmrs/esm-framework';
+import { useVisit } from './visit.resource';
+import { type Encounter } from '../types';
+import { formatDatetime, formatTime, parseDate } from '@openmrs/esm-framework';
 import EncounterList from './visits-components/encounter-list.component';
 import VisitSummary from './visits-components/visit-summary.component';
 import styles from './visit-detail-overview.scss';
