+++ packages/esm-active-visits-app/src/visits-summary/visit.resource.ts
@@ -1,139 +1,5 @@
 import useSWR from 'swr';
-import { openmrsFetch, restBaseUrl, type OpenmrsResource, type Visit } from '@openmrs/esm-framework';
-
-export interface Encounter {
-  uuid: string;
-  encounterDateTime: string;
-  encounterProviders: Array<{
-    uuid: string;
-    display: string;
-    encounterRole: {
-      uuid: string;
-      display: string;
-    };
-    provider: {
-      uuid: string;
-      person: {
-        uuid: string;
-        display: string;
-      };
-    };
-  }>;
-  encounterType: {
-    uuid: string;
-    display: string;
-  };
-  obs: Array<Observation>;
-  orders: Array<Order>;
-}
-
-export interface EncounterProvider {
-  uuid: string;
-  display: string;
-  encounterRole: {
-    uuid: string;
-    display: string;
-  };
-  provider: {
-    uuid: string;
-    person: {
-      uuid: string;
-      display: string;
-    };
-  };
-}
-
-export interface Observation {
-  uuid: string;
-  concept: {
-    uuid: string;
-    display: string;
-    conceptClass: {
-      uuid: string;
-      display: string;
-    };
-  };
-  display: string;
-  groupMembers: null | Array<{
-    uuid: string;
-    concept: {
-      uuid: string;
-      display: string;
-    };
-    value: {
-      uuid: string;
-      display: string;
-    };
-  }>;
-  value: any;
-  obsDatetime: string;
-}
-
-export interface Order {
-  uuid: string;
-  dateActivated: string;
-  dateStopped?: Date | null;
-  dose: number;
-  dosingInstructions: string | null;
-  dosingType?: 'org.openmrs.FreeTextDosingInstructions' | 'org.openmrs.SimpleDosingInstructions';
-  doseUnits: {
-    uuid: string;
-    display: string;
-  };
-  drug: {
-    uuid: string;
-    name: string;
-    strength: string;
-    display: string;
-  };
-  duration: number;
-  durationUnits: {
-    uuid: string;
-    display: string;
-  };
-  frequency: {
-    uuid: string;
-    display: string;
-  };
-  numRefills: number;
-  orderNumber: string;
-  orderReason: string | null;
-  orderReasonNonCoded: string | null;
-  orderer: {
-    uuid: string;
-    person: {
-      uuid: string;
-      display: string;
-    };
-  };
-  orderType: {
-    uuid: string;
-    display: string;
-  };
-  route: {
-    uuid: string;
-    display: string;
-  };
-  quantity: number;
-  quantityUnits: OpenmrsResource;
-}
-
-export interface Note {
-  note: string;
-  provider: {
-    name: string;
-    role: string;
-  };
-  time: string;
-}
-
-export interface OrderItem {
-  order: Order;
-  provider: {
-    name: string;
-    role: string;
-  };
-}
+import { openmrsFetch, restBaseUrl, type Visit } from '@openmrs/esm-framework';
 
 export function useVisit(visitUuid: string) {
   const customRepresentation =
